driver core: Add device "authorized" attribute support
In Confidential computing (CC) platforms (like AMD SEV and Intel TDX),
the host is an untrusted entity and hence the guest will guard against
any attacks from the host. For example the host can emulate a device
with incorrect PCI IDs and make the VM guest load a driver with
security holes. To harden against such attacks, guests can selectively
authorize the devices to bind the drivers. Currently bus drivers
like "USB" or "Thunderbolt" also has similar requirements and
implements a custom version of device authorization. Since we have
generic device authorization requirement for CC platforms, add the
support to the driver core and allow it to be used by all subsystems or
buses.

Use "authorized" device property to track and control the authorization
of the device. Acceptable property values are,

1 - Authorized
0 - Not authorized

Acceptable values in the resulting sysfs file are,

"on" - Authorized
"off" - Not Authorized"

By design only authorized devices are allowed to bind drivers. By
default all devices are set in authorized state (in
device_initalize()).

For cases like Thunderbolt bus where the "authorized" state is non
boolean, it does not fit well with the general device "authorization"
framework. So to allow such bus devices to opt out of core "authorized"
support, add "authorizable" flag in struct device. To opt out, bus
drivers can call dev_set_authorizable(dev, false) before device_add()
or device_register().

If the device is not authorizable then driver core will skip the
authorized status check during the driver probe operation and will also
not create authorized sysfs file for it.

This patch adds the core support. Modifications required in the USB and
Thunderbolt drivers to adapt to this support will be added by the
follow on patches.

Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>