driver core: Add command line option to initialize authorized attribute
Authorized device attribute is used to authorize or deauthorize
the driver probe of the given device. Currently this attribute
is initialized to "true" (allow all) by default.

But for confidential computing platforms like TDX guest, in which the
host is an untrusted entity, it has a requirement to disable all
devices by default and allow only a set of acceptable devices. Support
to allow only an acceptable list of devices to be added by the follow
on patches. As a first step in adding this support, add a command line
option "dev.authorize.all" to allow platforms to change the default
authorized status.

Also note that USB and Thunderbolt both override this initial value
in their respective device initializations so this is not a regression
for those buses.

Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>