RISC-V: Add support for Ssqosid extension
The extension adds a new sqoscfg csr that is used to set Monitor and
Resource Control IDs.
They're used by a cache/memory controller to tag transactions,
monitor them and possibly enforce resource usage limits.
The IDs are stored in the task struct and updated during context switch.
In order to save some cycles a per cpu struct with the current state
has been created, it's used to check if an update is needed.
The resoning behind this is that an access to L1D hot memory should
be several times faster than a csr read.
Also in the case of virtualization accesses to this csr are trapped
in the hypervisor.

Signed-off-by: Kornel Dulęba <mindal@semihalf.com>