RISC-V: KVM: Add AIA support for TVMs
The AIA support for TVMs are split between the host and the TSM.
While the host allocates the vsfile, the TSM controls the gstage mapping
and any updates to it. The host must not be able to inject interrupt
to a TVM. Thus, the interrupt injection has to happen via the TSM only
for interrupts allowed by the guest.

The swfile is kind of useless for the TVMs as well as the TVMs only
work for HW_ACCEL mode. The swfile allocation is kept as is to avoid
further bifurcation of the code.

Signed-off-by: Atish Patra <atishp@rivosinc.com>