asm/io.h: Add iounmap_driver_hardened
This function is for unmapping memory that was shared with a hypervisor
in a confidential guest. If the architecture doesn't implement it it's
just iounmap.

Note this function is not supposed to be used directly by drivers.
But it will be used implicitly by pci_iounmap -- based on the
authorized attribute -- and also by the PCI subsystem.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>