RISC-V: KVM: Add UABI to support static memory region attestation
To initialize a TVM, a TSM must ensure that all the static memory regions
that contains the device tree or the kernel image or initrd for the TVM
attested. This information is not usually present with the host and only
VMM is aware of these.

Introduce an new ioctl which is part of the uABI to support this.

Signed-off-by: Atish Patra <atishp@rivosinc.com>