kernel: Match preferred console first.
For cases where cmdline has "console=hvc0 console =hvc1"
the expected flow is that hvc1 will take precedence but
given hvc console has index set to -1, the current logic
will try to assign the first console that matches which
is hvc0.

This change makes sure if the index is -1 for a console
make sure to check preferred console first.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>