riscv: pmu: Add support for Control Transfer Records Ext.
This adds support for CTR Ext defined in [0]. The extension
allows to records a maximum for 256 last branch records.

CTR extension depends on csrind and Sscofpmf extensions.

In the current model, CTR integrates into existing
riscv_pmu_sbi.c and we sample CTR entries on counter OVF
interrupts for the events with BRANCH_STACK enabled.

[0]: https://github.com/riscv/riscv-control-transfer-records

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>